+++ components/form/FileImage.js
@@ -1,5 +1,3 @@
-import 'isomorphic-fetch';
-
 import React from 'react';
 import PropTypes from 'prop-types';