+++ components/tools/related-tools/initial-state.js
@@ -1,5 +1,0 @@
-export default {
-  list: [],
-  loading: true,
-  error: null,
-};