+++ layout/search/search-actions.js
@@ -1,4 +1,1 @@
-import 'isomorphic-fetch';
-
-// Actions
 export * from 'components/search-results/actions';